Add JavaDoc comments, use setJMenuBar over add
This commit adds JavaDoc comments to all of the methods in the class and parts of the code. It also cleans up misc. areas of the code to make them more friendly for comments.

Additionally, the add() method for the JMenuBar was replaced with the setJMenuBar() method, giving more flexibility within the frame for adding more things.