Add JavaDoc comments, make MouseListener to inner class MouseAdapter
This commit adds JavaDoc comments to all methods, variables, and definitive logic that might need an explanation. These documentation changes intend to make the code easier to read and understand not only for the immediate sense and for the grade, but also for revisiting later, such as in a final project context...

Additionally, this class originally implemented a MouseListener for handling mouse events, but that required defining several methods that are not used for our implementation of this project. To resolve this, I created an inner class that extends MouseAdapter, granting us the flexibility to only use methods that we choose to use.