Implement checkForMove method, map out where game ends, begin logic for piece conversion
This is proving quite tricky. I've begun implementing preliminary logic for how the game will scan the board to find neighboring pieces (and thereby set that piece to status 1). I know at what point the game ends, but there is no way to handle the end of the game yet (this will be easy).

There is currently an IndexOutOfBounds exception when the board is first scanned for moves, but I need to play around with it more tomorrow. This is where I'm leaving off for tonight.